refactor: move `Results` to a separate file
This commit moves the `Results` type and its
functions to a separate file. This is done to make
the code more readable and maintainable.
